+++ src/main/resources/counter-application.conf
@@ -16,7 +16,7 @@
 include "general"
 
 postgres-journal {
-  dao = "com.github.dnvriend.dao.CounterJournalDao"
+  dao = "com.github.mkubala.dao.CounterJournalDao"
   slick = ${slick}
 
   tables.journal.schemaName = "flat"
