+++ src/main/resources/person-application.conf
@@ -46,24 +46,24 @@ akka {
     }
 
     serializers {
-      createPerson = "com.github.dnvriend.serializer.CreatePersonSerializer"
-      changeFirstName = "com.github.dnvriend.serializer.ChangeFirstNameSerializer"
-      changeLastName = "com.github.dnvriend.serializer.ChangeLastNameSerializer"
+      createPerson = "com.github.mkubala.serializer.CreatePersonSerializer"
+      changeFirstName = "com.github.mkubala.serializer.ChangeFirstNameSerializer"
+      changeLastName = "com.github.mkubala.serializer.ChangeLastNameSerializer"
 
-      personCreated = "com.github.dnvriend.serializer.PersonCreatedSerializer"
-      firstNameChanged = "com.github.dnvriend.serializer.FirstNameChangedSerializer"
-      lastNameChanged = "com.github.dnvriend.serializer.LastNameChangedSerializer"
+      personCreated = "com.github.mkubala.serializer.PersonCreatedSerializer"
+      firstNameChanged = "com.github.mkubala.serializer.FirstNameChangedSerializer"
+      lastNameChanged = "com.github.mkubala.serializer.LastNameChangedSerializer"
     }
 
     serialization-bindings {
-      "com.github.dnvriend.Person$CreatePerson" = createPerson
-      "com.github.dnvriend.Person$ChangeFirstName" = changeFirstName
-      "com.github.dnvriend.Person$ChangeLastName" = changeLastName
+      "com.github.mkubala.Person$CreatePerson" = createPerson
+      "com.github.mkubala.Person$ChangeFirstName" = changeFirstName
+      "com.github.mkubala.Person$ChangeLastName" = changeLastName
 
-      "com.github.dnvriend.data.Command.PBCreatePerson" = createPerson
-      "com.github.dnvriend.data.Event.PBPersonCreated" = personCreated
-      "com.github.dnvriend.data.Event.PBFirstNameChanged" = firstNameChanged
-      "com.github.dnvriend.data.Event.PBLastNameChanged" = lastNameChanged
+      "com.github.mkubala.data.Command.PBCreatePerson" = createPerson
+      "com.github.mkubala.data.Event.PBPersonCreated" = personCreated
+      "com.github.mkubala.data.Event.PBFirstNameChanged" = firstNameChanged
+      "com.github.mkubala.data.Event.PBLastNameChanged" = lastNameChanged
     }
   }
 }
@@ -81,16 +81,16 @@ postgres-journal {
   // akka persistence can convert domain messages to messages specific for
   // the plugin, like eg. tagging an event
   event-adapters {
-    tagging = "com.github.dnvriend.adapter.TaggingEventAdapter"
+    tagging = "com.github.mkubala.adapter.TaggingEventAdapter"
   }
 
   event-adapter-bindings {
-    "com.github.dnvriend.data.Event.PBPersonCreated" = tagging
-    "com.github.dnvriend.data.Event.PBFirstNameChanged" = tagging
-    "com.github.dnvriend.data.Event.PBLastNameChanged" = tagging
-    "com.github.dnvriend.Person$PersonCreated" = tagging
-    "com.github.dnvriend.Person$FirstNameChanged" = tagging
-    "com.github.dnvriend.Person$LastNameChanged" = tagging
+    "com.github.mkubala.data.Event.PBPersonCreated" = tagging
+    "com.github.mkubala.data.Event.PBFirstNameChanged" = tagging
+    "com.github.mkubala.data.Event.PBLastNameChanged" = tagging
+    "com.github.mkubala.Person$PersonCreated" = tagging
+    "com.github.mkubala.Person$FirstNameChanged" = tagging
+    "com.github.mkubala.Person$LastNameChanged" = tagging
   }
 }
