+++ src/main/resources/pet-application.conf
@@ -20,7 +20,7 @@ akka {
 
   actor {
     serialization-bindings {
-      "com.github.dnvriend.domain.PetDomain$Pet" = jackson-cbor
+      "com.github.mkubala.domain.PetDomain$Pet" = jackson-cbor
     }
   }
 }
