+++ src/main/resources/wrapper-application.conf
@@ -17,16 +17,16 @@ include "general"
 
 # the postgres-journal in use
 postgres-journal {
-  dao = "com.github.dnvriend.dao.WrapperByteArrayJournalDao"
+  dao = "com.github.mkubala.dao.WrapperByteArrayJournalDao"
 
   slick = ${slick}
 
   event-adapters {
-    wrapper = "com.github.dnvriend.adapter.WrapperEventAdapter"
+    wrapper = "com.github.mkubala.adapter.WrapperEventAdapter"
   }
 
   event-adapter-bindings {
-    "com.github.dnvriend.adapter.Wrapper" = wrapper
+    "com.github.mkubala.adapter.Wrapper" = wrapper
   }
 }
