+++ src/main/scala/com/github/mkubala/serializer/ChangeFirstNameSerializer.scala
@@ -15,20 +15,21 @@
  * limitations under the License.
  */
 
-package com.github.dnvriend.serializer
+package com.github.mkubala.serializer
 
 import akka.serialization.SerializerWithStringManifest
-import com.github.dnvriend.data.Event.PBFirstNameChanged
+import com.github.mkubala.Person.ChangeFirstName
+import com.github.mkubala.data.Command.PBChangeFirstName
 
 /**
  * Converts FirstName Google Protobuf Message
  * to byte array and back
  */
-class FirstNameChangedSerializer extends SerializerWithStringManifest {
+class ChangeFirstNameSerializer extends SerializerWithStringManifest {
 
-  override def identifier: Int = 103
+  override def identifier: Int = 100
 
-  final val Manifest = classOf[PBFirstNameChanged].getName
+  final val Manifest = classOf[ChangeFirstName].getName
 
   override def manifest(o: AnyRef): String = o.getClass.getName
 
@@ -36,14 +37,16 @@ class FirstNameChangedSerializer extends SerializerWithStringManifest {
    * Unmarshal to the data model
    */
   override def fromBinary(bytes: Array[Byte], manifest: String): AnyRef =
-    if (Manifest == manifest) PBFirstNameChanged.parseFrom(bytes)
-    else throw new IllegalArgumentException("Unable to handle manifest: " + manifest)
+    if (Manifest == manifest) {
+      val PBChangeFirstName(firstName, timestamp) = PBChangeFirstName.parseFrom(bytes)
+      ChangeFirstName(firstName, timestamp)
+    } else throw new IllegalArgumentException("Unable to handle manifest: " + manifest)
 
   /**
    * Marshal the data model to bytes
    */
   override def toBinary(o: AnyRef): Array[Byte] = o match {
-    case e: PBFirstNameChanged ⇒ e.toByteArray
-    case _                     ⇒ throw new IllegalStateException("Cannot serialize: " + o.getClass.getName)
+    case ChangeFirstName(firstName, timestamp) ⇒ PBChangeFirstName(firstName, timestamp).toByteArray
+    case _                                     ⇒ throw new IllegalStateException("Cannot serialize: " + o.getClass.getName)
   }
 }
