+++ src/main/scala/com/github/mkubala/serializer/FirstNameChangedSerializer.scala
@@ -15,29 +15,35 @@
  * limitations under the License.
  */
 
-package com.github.dnvriend.serializer
+package com.github.mkubala.serializer
 
 import akka.serialization.SerializerWithStringManifest
-import com.github.dnvriend.data.Event.{ PBLastNameChanged, PBFirstNameChanged }
+import com.github.mkubala.data.Event.PBFirstNameChanged
 
 /**
- * Converts LastNameChanged Google Protobuf Message
- * to byte array and back.
+ * Converts FirstName Google Protobuf Message
+ * to byte array and back
  */
-class LastNameChangedSerializer extends SerializerWithStringManifest {
+class FirstNameChangedSerializer extends SerializerWithStringManifest {
 
-  override def identifier: Int = 104
+  override def identifier: Int = 103
 
-  final val Manifest = classOf[PBLastNameChanged].getName
+  final val Manifest = classOf[PBFirstNameChanged].getName
 
   override def manifest(o: AnyRef): String = o.getClass.getName
 
+  /**
+   * Unmarshal to the data model
+   */
   override def fromBinary(bytes: Array[Byte], manifest: String): AnyRef =
-    if (Manifest == manifest) PBLastNameChanged.parseFrom(bytes)
+    if (Manifest == manifest) PBFirstNameChanged.parseFrom(bytes)
     else throw new IllegalArgumentException("Unable to handle manifest: " + manifest)
 
+  /**
+   * Marshal the data model to bytes
+   */
   override def toBinary(o: AnyRef): Array[Byte] = o match {
-    case e: PBLastNameChanged ⇒ e.toByteArray
-    case _                    ⇒ throw new IllegalStateException("Cannot serialize: " + o.getClass.getName)
+    case e: PBFirstNameChanged ⇒ e.toByteArray
+    case _                     ⇒ throw new IllegalStateException("Cannot serialize: " + o.getClass.getName)
   }
 }
