+++ src/main/scala/com/github/mkubala/serializer/LastNameChangedSerializer.scala
@@ -15,29 +15,29 @@
  * limitations under the License.
  */
 
-package com.github.dnvriend.serializer
+package com.github.mkubala.serializer
 
 import akka.serialization.SerializerWithStringManifest
-import com.github.dnvriend.data.Event.PBPersonCreated
+import com.github.mkubala.data.Event.{ PBLastNameChanged, PBFirstNameChanged }
 
 /**
- * Converts PersonCreated Google Protobuf Message
- * to byte array and back
+ * Converts LastNameChanged Google Protobuf Message
+ * to byte array and back.
  */
-class PersonCreatedSerializer extends SerializerWithStringManifest {
+class LastNameChangedSerializer extends SerializerWithStringManifest {
 
-  override def identifier: Int = 105
+  override def identifier: Int = 104
 
-  final val Manifest = classOf[PBPersonCreated].getName
+  final val Manifest = classOf[PBLastNameChanged].getName
 
   override def manifest(o: AnyRef): String = o.getClass.getName
 
   override def fromBinary(bytes: Array[Byte], manifest: String): AnyRef =
-    if (Manifest == manifest) PBPersonCreated.parseFrom(bytes)
+    if (Manifest == manifest) PBLastNameChanged.parseFrom(bytes)
     else throw new IllegalArgumentException("Unable to handle manifest: " + manifest)
 
   override def toBinary(o: AnyRef): Array[Byte] = o match {
-    case e: PBPersonCreated ⇒ e.toByteArray
-    case _                  ⇒ throw new IllegalStateException("Cannot serialize: " + o.getClass.getName)
+    case e: PBLastNameChanged ⇒ e.toByteArray
+    case _                    ⇒ throw new IllegalStateException("Cannot serialize: " + o.getClass.getName)
   }
 }
