+++ src/main/scala/com/github/mkubala/serializer/PersonCreatedSerializer.scala
@@ -15,38 +15,29 @@
  * limitations under the License.
  */
 
-package com.github.dnvriend.serializer
+package com.github.mkubala.serializer
 
 import akka.serialization.SerializerWithStringManifest
-import com.github.dnvriend.Person.ChangeLastName
-import com.github.dnvriend.data.Command.PBChangeLastName
+import com.github.mkubala.data.Event.PBPersonCreated
 
 /**
- * Converts FirstName Google Protobuf Message
+ * Converts PersonCreated Google Protobuf Message
  * to byte array and back
  */
-class ChangeLastNameSerializer extends SerializerWithStringManifest {
+class PersonCreatedSerializer extends SerializerWithStringManifest {
 
-  override def identifier: Int = 101
+  override def identifier: Int = 105
 
-  final val Manifest = classOf[ChangeLastName].getName
+  final val Manifest = classOf[PBPersonCreated].getName
 
   override def manifest(o: AnyRef): String = o.getClass.getName
 
-  /**
-   * Unmarshal to the data model
-   */
   override def fromBinary(bytes: Array[Byte], manifest: String): AnyRef =
-    if (Manifest == manifest) {
-      val PBChangeLastName(lastName, timestamp) = PBChangeLastName.parseFrom(bytes)
-      ChangeLastName(lastName, timestamp)
-    } else throw new IllegalArgumentException("Unable to handle manifest: " + manifest)
+    if (Manifest == manifest) PBPersonCreated.parseFrom(bytes)
+    else throw new IllegalArgumentException("Unable to handle manifest: " + manifest)
 
-  /**
-   * Marshal the data model to bytes
-   */
   override def toBinary(o: AnyRef): Array[Byte] = o match {
-    case ChangeLastName(lastName, timestamp) ⇒ PBChangeLastName(lastName, timestamp).toByteArray
-    case _                                   ⇒ throw new IllegalStateException("Cannot serialize: " + o.getClass.getName)
+    case e: PBPersonCreated ⇒ e.toByteArray
+    case _                  ⇒ throw new IllegalStateException("Cannot serialize: " + o.getClass.getName)
   }
 }
